requirements.txt
Django==6.0.1
djangorestframework==3.16.1
drf-spectacular==0.29.0
channels==4.3.2
daphne==4.2.1
djangorestframework-simplejwt==5.5.1
django-ckeditor-5==0.2.19
pillow==12.1.1
